+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added capability of regex matching audio device names https://github.com/GrandOrgue/grandorgue/issues/1265
 - Added a checkbox-based Stops dialog (Audio/Midi -> Stops) https://github.com/GrandOrgue/grandorgue/issues/1816
 - Fixed not compatibilyty between polyphase interpolation and loseless compression https://github.com/GrandOrgue/grandorgue/issues/710
 - Fixed not sounding of pipe after an enclosure has closed to zero and then opened https://github.com/GrandOrgue/grandorgue/issues/1813